src/python_prototype/vm.py
# vm.py

from enum import Enum, auto
from typing import List, Any, Dict

class OpCode(Enum):
    LOAD_CONST = auto()
    ADD = auto()
    RETURN_VALUE = auto()
    ROT_THREE = auto()
    MAKE_FUNCTION = auto()
    GET_ITER = auto()
    POP_JUMP_IF_FALSE = auto()
    STORE_NAME = auto()
    LOAD_NAME = auto()
    CALL_FUNCTION = auto()
    BINARY_ADD = auto()
    BINARY_SUBTRACT = auto()
    BINARY_MULTIPLY = auto()
    BINARY_DIVIDE = auto()
    COMPARE_OP = auto()
    # Add other opcodes as needed

class Instruction:
    def __init__(self, opcode: OpCode, arg: Any = None):
        self.opcode = opcode
        self.arg = arg

class VirtualMachine:
    def __init__(self):
        self.stack: List[Any] = []
        self.locals: Dict[str, Any] = {}
        self.globals: Dict[str, Any] = {}
        self.frame = None

    def execute(self, instructions: List[Instruction]) -> Any:
        self.stack = []
        for instruction in instructions:
            if instruction.opcode == OpCode.LOAD_CONST:
                self.stack.append(instruction.arg)
            elif instruction.opcode == OpCode.STORE_NAME:
                self.locals[instruction.arg] = self.stack.pop()
            elif instruction.opcode == OpCode.LOAD_NAME:
                self.stack.append(self.locals[instruction.arg])
            elif instruction.opcode == OpCode.BINARY_ADD:
                b = self.stack.pop()
                a = self.stack.pop()
                self.stack.append(a + b)
            elif instruction.opcode == OpCode.MAKE_FUNCTION:
                code = self.stack.pop()
                self.stack.append(lambda: self.execute(code))
            elif instruction.opcode == OpCode.GET_ITER:
                self.stack.append(iter(self.stack.pop()))
            elif instruction.opcode == OpCode.POP_JUMP_IF_FALSE:
                if not self.stack.pop():
                    return instruction.arg
            elif instruction.opcode == OpCode.RETURN_VALUE:
                return self.stack.pop() if self.stack else None

        return None
